Expose the new leader schedule functionality from the bank.
