+++ runtime/Cargo.toml
@@ -15,6 +15,7 @@ fnv = "1.0.6"
 hashbrown = "0.1.8"
 log = "0.4.2"
 rand = "0.6.5"
+rand_chacha = "0.1.1"
 serde = "1.0.88"
 serde_derive = "1.0.88"
 serde_json = "1.0.38"
