+++ runtime/src/lib.rs
@@ -2,6 +2,7 @@ mod accounts;
 pub mod bank;
 pub mod bloom;
 mod last_id_queue;
+mod leader_schedule;
 mod runtime;
 mod status_cache;
